requirements.txt
Flask==2.3.3
torch==2.0.1
torchvision==0.15.2
numpy==1.26.4
opencv-python==4.8.1.78
Pillow==9.5.0
